Add simple installation logic, get full bundle compiling
Note that we still do not correctly model priorities here, and we also do not have api tests.

The C API is also woefully under-represented so far, but ultimately, we have the core testing stuff set up.

We still need to add validation for the installation failures and improved logging to ensure reasonable visibility of the hook graph (in some pretty-ish way)

And also we want to have tests that run on arm64 too :-)